backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
groq
langchain
langchain-community
datasets
python-dotenv
requests
